Rename to promiseMap.ts
